@@ -1,11 +1,12 @@
 <script lang="ts">
 	import { createEventDispatcher } from 'svelte'
+	import classnames from 'classnames'
+	import { Spinner } from '$lib/components'
+	import { fly } from 'svelte/transition'
 
 	export let type: 'submit' | 'button' | 'reset' | undefined = undefined
 	export let href: string | undefined = undefined
-	// TODO: Need to make primary and secondary one prop as they can't be both
-	export let primary = true
-	export let secondary = false
+	export let variant: 'primary' | 'secondary' = 'primary'
 	export let fluid = false
 	export let style = ''
 	export let small = false
@@ -14,7 +15,16 @@
 	export let value: string | null = null
 	export let id: string | null = null
 	export let form: string | null = null
+	export let delayed = false
+	export let delayedReason: string | null = null
 
+	const classes = classnames({
+		primary: variant === 'primary',
+		secondary: variant === 'secondary',
+		fluid,
+		small,
+		danger,
+	})
 	const element = href ? 'a' : 'button'
 	const role = element === 'a' ? 'link' : 'button'
 
@@ -26,32 +36,43 @@
 
 <svelte:element
 	this={element}
-	{type}
-	{href}
 	{role}
 	{style}
 	{name}
 	{value}
 	{id}
 	{form}
-	class="base"
-	class:primary
-	class:fluid
-	class:secondary
-	class:small
-	class:danger
+	{type}
+	{href}
+	class="base {classes}"
 	on:click={onClick}
+	disabled={delayed}
+	aria-disabled={delayed ? 'true' : undefined}
 >
-	<slot />
+	{#if delayed}
+		<span transition:fly={{ y: 20, duration: 200 }} class="delayed-text">
+			<Spinner
+				color={variant === 'primary'
+					? 'var(--palette-primary)'
+					: 'var(--palette-secondary)'}
+			/>
+			{delayedReason}
+		</span>
+	{:else}
+		<span class="delayed-text">
+			<slot />
+		</span>
+	{/if}
 </svelte:element>
 
 <style>
 	.base {
 		--padding: var(--space-xs);
 		--font-size: var(--type-step-0);
+		--gap: var(--space-3xs);
 		display: inline-flex;
 		align-items: center;
-		gap: var(--space-3xs);
+		gap: var(--gap);
 		padding: var(--padding);
 		background: var(--background);
 		border-radius: var(--border-radius);
@@ -97,5 +118,15 @@
 	.fluid {
 		--width: 100%;
 		justify-content: center;
+	}
+
+	.disabled {
+		cursor: not-allowed;
+	}
+
+	.delayed-text {
+		display: flex;
+		align-items: center;
+		gap: var(--space-2xs);
 	}
 </style>
