@@ -2,13 +2,14 @@ export { default as AvatarMenu } from '$lib/components/AvatarMenu/AvatarMenu.sve
 export { default as Button } from '$lib/components/Button/Button.svelte'
 export { default as EditNote } from '$lib/components/EditNote/EditNote.svelte'
 export { default as HoneypotInputs } from '$lib/components/HoneypotInputs/HoneypotInputs.svelte'
-export { default as Icon } from '$lib/components/icons/Icon.svelte'
+// export { default as Icon } from '$lib/components/icons/figureOutSpriteIcons/Icon.svelte'
+export * from '$lib/components/icons'
 export { default as ImageEditor } from '$lib/components/ImageEditor/ImageEditor.svelte'
 export { Input, TextArea, FormGroup } from '$lib/components/FormElements'
 export { default as Navbar } from '$lib/components/Navbar/Navbar.svelte'
 export { default as NoteInfoBar } from '$lib/components/NoteInfoBar/NoteInfoBar.svelte'
 export { default as ProviderConnectionForm } from '$lib/components/ProviderConnectionForm/ProviderConnectionForm.svelte'
 export { default as Searchbar } from '$lib/components/Search/Searchbar.svelte'
+export { default as Spinner } from '$lib/components/Spinner/Spinner.svelte'
 export { default as Toast } from '$lib/components/Toast/Toast.svelte'
 export { default as ValidationErrors } from '$lib/components/ValidationErrors/ValidationErrors.svelte'
-export * from '$lib/components/Search/search.types'
