@@ -34,7 +34,7 @@
 			type="text"
 		/>
 	</FormGroup>
-	<Button fluid type="submit" secondary>Submit</Button>
+	<Button fluid type="submit" variant="secondary">Submit</Button>
 	<ValidationErrors errorId={formId} errors={$errors._errors} />
 </form>
 
