@@ -50,7 +50,7 @@
 	</FormGroup>
 	<div class="button-wrapper">
 		<a class="link-button" href=".">Cancel</a>
-		<Button secondary type="submit">Change Password</Button>
+		<Button variant="secondary" type="submit">Change Password</Button>
 	</div>
 </form>
 
