@@ -1,46 +1,56 @@
 <script lang="ts">
 	import { enhance } from '$app/forms'
-	import Button from '$lib/components/Button/Button.svelte'
-	import Timer from 'virtual:icons/radix-icons/timer'
-	import Eraser from 'virtual:icons/radix-icons/eraser'
-	import Pencil2 from 'virtual:icons/radix-icons/pencil2'
 	import { page } from '$app/stores'
-	import { NoteInfoBar } from '$lib/components'
+	import {
+		Button,
+		NoteInfoBar,
+		Eraser,
+		Pencil2,
+		Timer,
+		Spinner,
+	} from '$lib/components'
 	import { getNoteImgSrc } from '$lib/utils/misc.js'
 
+	// export let form
 	export let data
 </script>
 
-<article class="article">
-	<h2 class="heading">{data.note.title}</h2>
-	<ul class="image-list">
-		{#each data.note.images as image}
-			<li>
-				<a href={getNoteImgSrc(image.id)}>
-					<img src={getNoteImgSrc(image.id)} alt={image.altText ?? ''} />
-				</a>
-			</li>
+{#await data.note}
+	<Spinner />
+{:then}
+	<article class="article">
+		<h2 class="heading">{data.note.title}</h2>
+		<ul class="image-list">
+			{#each data.note.images as image}
+				<li>
+					<a href={getNoteImgSrc(image.id)}>
+						<img src={getNoteImgSrc(image.id)} alt={image.altText ?? ''} />
+					</a>
+				</li>
+			{/each}
+		</ul>
+
+		{#each data.note.content as paragraph}
+			<p>{paragraph}</p>
 		{/each}
-	</ul>
+	</article>
 
-	{#each data.note.content as paragraph}
-		<p>{paragraph}</p>
-	{/each}
-</article>
-
-{#if data.isOwner}
-	<NoteInfoBar>
-		<span class="time-since-update"><Timer />{data.timeSinceUpdate}</span>
-		<div class="buttons">
-			<Button small secondary href="{$page.params.noteId}/edit">
-				<Pencil2 /> Edit
-			</Button>
-			<form method="POST" use:enhance>
-				<Button small danger type="submit"><Eraser /> Delete</Button>
-			</form>
-		</div>
-	</NoteInfoBar>
-{/if}
+	{#if data.isOwner}
+		<NoteInfoBar>
+			<span class="time-since-update"><Timer />{data.timeSinceUpdate}</span>
+			<div class="buttons">
+				<Button small variant="secondary" href="{$page.params.noteId}/edit">
+					<Pencil2 /> Edit
+				</Button>
+				<form method="POST" use:enhance>
+					<Button small danger type="submit"><Eraser /> Delete</Button>
+				</form>
+			</div>
+		</NoteInfoBar>
+	{/if}
+{:catch error}
+	<p>{error.message}</p>
+{/await}
 
 <style>
 	.article {
